refactor(dynamite,dynamite_runtime,nextcloud)!: generated clients no longer implement http.Client
