+++ packages/dynamite/dynamite/example/lib/petstore.openapi.dart
@@ -39,18 +39,14 @@ class $Client extends _i1.DynamiteClient {
   /// Creates a new `DynamiteClient` for untagged requests.
   $Client(
     super.baseURL, {
-    super.baseHeaders,
     super.httpClient,
-    super.cookieJar,
   });
 
   /// Creates a new [$Client] from another [client].
   $Client.fromClient(_i1.DynamiteClient client)
       : super(
           client.baseURL,
-          baseHeaders: client.baseHeaders,
           httpClient: client.httpClient,
-          cookieJar: client.cookieJar,
           authentications: client.authentications,
         );
 
@@ -130,7 +126,7 @@ class $Client extends _i1.DynamiteClient {
       tags: tags,
       limit: limit,
     );
-    final _response = await send(_request);
+    final _response = await httpClient.send(_request);
 
     final _serializer = $findPets_Serializer();
     final _rawResponse = await _i1.ResponseConverter<BuiltList<Pet>, void>(_serializer).convert(_response);
@@ -185,7 +181,7 @@ class $Client extends _i1.DynamiteClient {
     final _request = $addPet_Request(
       newPet: newPet,
     );
-    final _response = await send(_request);
+    final _response = await httpClient.send(_request);
 
     final _serializer = $addPet_Serializer();
     final _rawResponse = await _i1.ResponseConverter<Pet, void>(_serializer).convert(_response);
@@ -248,7 +244,7 @@ class $Client extends _i1.DynamiteClient {
     final _request = $findPetById_Request(
       id: id,
     );
-    final _response = await send(_request);
+    final _response = await httpClient.send(_request);
 
     final _serializer = $findPetById_Serializer();
     final _rawResponse = await _i1.ResponseConverter<Pet, void>(_serializer).convert(_response);
@@ -310,7 +306,7 @@ class $Client extends _i1.DynamiteClient {
     final _request = $deletePet_Request(
       id: id,
     );
-    final _response = await send(_request);
+    final _response = await httpClient.send(_request);
 
     final _serializer = $deletePet_Serializer();
     final _rawResponse = await _i1.ResponseConverter<void, void>(_serializer).convert(_response);
