+++ packages/dynamite/dynamite_end_to_end_test/test/request_body_test.dart
@@ -38,7 +38,7 @@ void main() async {
         return Response('{}', 200);
       }),
     );
-    await client.$post();
+    await client.post();
 
     // with body
     client = $Client(
@@ -48,6 +48,6 @@ void main() async {
         return Response('{}', 200);
       }),
     );
-    await client.$post(string: 'value');
+    await client.post(string: 'value');
   });
 }
