+++ packages/dynamite/dynamite_runtime/lib/src/client/client.dart
@@ -1,12 +1,8 @@
-import 'dart:async';
-
-import 'package:cookie_jar/cookie_jar.dart';
 import 'package:dynamite_runtime/src/client/authentication.dart';
 import 'package:dynamite_runtime/src/client/exception.dart';
 import 'package:dynamite_runtime/src/client/response.dart';
 import 'package:dynamite_runtime/src/utils/uri.dart';
 import 'package:http/http.dart' as http;
-import 'package:meta/meta.dart';
 
 /// A client for making network requests.
 ///
@@ -15,21 +11,20 @@ import 'package:meta/meta.dart';
 ///   * [DynamiteRawResponse] as the raw response that can be serialized.
 ///   * [DynamiteApiException] as the exception that can be thrown in operations
 ///   * [DynamiteAuthentication] for providing authentication methods.
-class DynamiteClient with http.BaseClient {
+abstract class DynamiteClient {
   /// Creates a new dynamite network client.
   ///
   /// If [httpClient] is not provided a default one will be created.
   /// The [baseURL] will be normalized, removing any trailing `/`.
+  /// It is an error if the [baseURL] contains any query parameters.
   DynamiteClient(
     Uri baseURL, {
-    this.baseHeaders,
     http.Client? httpClient,
-    this.cookieJar,
     this.authentications,
   })  : httpClient = httpClient ?? http.Client(),
         baseURL = baseURL.normalizeEmptyPath() {
     if (baseURL.queryParametersAll.isNotEmpty) {
-      throw UnsupportedError('Dynamite can not work with a baseURL containing query parameters.');
+      throw ArgumentError.value(baseURL, 'baseURL', 'MUST NOT contain query parameters.');
     }
   }
 
@@ -39,61 +34,11 @@ class DynamiteClient with http.BaseClient {
   /// further information.
   final Uri baseURL;
 
-  /// The base headers added to each request.
-  final Map<String, String>? baseHeaders;
-
   /// The base http client.
-  @protected
   final http.Client httpClient;
-
-  /// The optional cookie jar to persist the response cookies.
-  final CookieJar? cookieJar;
 
   /// The available authentications for this client.
   ///
   /// The first one matching the required authentication type will be used.
   final List<DynamiteAuthentication>? authentications;
-
-  /// Sends an HTTP request and asynchronously returns the response.
-  ///
-  /// Cookies are persisted in the [cookieJar] and loaded for requests.
-  @override
-  Future<http.StreamedResponse> send(http.BaseRequest request) async {
-    if (cookieJar != null) {
-      final cookies = await cookieJar!.loadForRequest(request.url);
-      if (cookies.isNotEmpty) {
-        final buffer = StringBuffer();
-
-        for (final entry in cookies.indexed) {
-          final cookie = entry.$2;
-
-          buffer
-            ..write(cookie.name)
-            ..write('=')
-            ..write(cookie.value);
-
-          if (entry.$1 < cookies.length - 1) {
-            buffer.write('; ');
-          }
-        }
-
-        request.headers['cookie'] = buffer.toString();
-      }
-    }
-
-    // Do not overwrite request headers to avoid invalid requests.
-    baseHeaders?.forEach((key, value) {
-      request.headers.putIfAbsent(key, () => value);
-    });
-
-    final response = await httpClient.send(request);
-
-    final cookieHeader = response.headersSplitValues['set-cookie'];
-    if (cookieHeader != null && cookieJar != null) {
-      final cookies = cookieHeader.map(Cookie.fromSetCookieValue).toList();
-      await cookieJar!.saveFromResponse(request.url, cookies);
-    }
-
-    return response;
-  }
 }
