+++ packages/nextcloud/lib/src/api/core.openapi.dart
@@ -37,9 +37,7 @@ class $Client extends _i1.DynamiteClient {
   /// Creates a new `DynamiteClient` for untagged requests.
   $Client(
     super.baseURL, {
-    super.baseHeaders,
     super.httpClient,
-    super.cookieJar,
     super.authentications,
   });
 
@@ -47,9 +45,7 @@ class $Client extends _i1.DynamiteClient {
   $Client.fromClient(_i1.DynamiteClient client)
       : super(
           client.baseURL,
-          baseHeaders: client.baseHeaders,
           httpClient: client.httpClient,
-          cookieJar: client.cookieJar,
           authentications: client.authentications,
         );
 
@@ -134,7 +130,7 @@ class $Client extends _i1.DynamiteClient {
   ///  * [$getStatus_Serializer] for a converter to parse the `Response` from an executed request.
   Future<_i1.DynamiteResponse<Status, void>> getStatus() async {
     final _request = $getStatus_Request();
-    final _response = await send(_request);
+    final _response = await httpClient.send(_request);
 
     final _serializer = $getStatus_Serializer();
     final _rawResponse = await _i1.ResponseConverter<Status, void>(_serializer).convert(_response);
@@ -224,7 +220,7 @@ class $AppPasswordClient {
     final _request = $getAppPassword_Request(
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $getAppPassword_Serializer();
     final _rawResponse =
@@ -309,7 +305,7 @@ class $AppPasswordClient {
     final _request = $rotateAppPassword_Request(
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $rotateAppPassword_Serializer();
     final _rawResponse =
@@ -394,7 +390,7 @@ class $AppPasswordClient {
     final _request = $deleteAppPassword_Request(
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $deleteAppPassword_Serializer();
     final _rawResponse =
@@ -539,7 +535,7 @@ class $AutoCompleteClient {
       limit: limit,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $$get_Serializer();
     final _rawResponse =
@@ -639,7 +635,7 @@ class $AvatarClient {
       userId: userId,
       size: size,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $getAvatarDark_Serializer();
     final _rawResponse =
@@ -730,7 +726,7 @@ class $AvatarClient {
       userId: userId,
       size: size,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $getAvatar_Serializer();
     final _rawResponse =
@@ -816,7 +812,7 @@ class $ClientFlowLoginV2Client {
     final _request = $poll_Request(
       token: token,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $poll_Serializer();
     final _rawResponse = await _i1.ResponseConverter<LoginFlowV2Credentials, void>(_serializer).convert(_response);
@@ -880,7 +876,7 @@ class $ClientFlowLoginV2Client {
   ///  * [$init_Serializer] for a converter to parse the `Response` from an executed request.
   Future<_i1.DynamiteResponse<LoginFlowV2, void>> init() async {
     final _request = $init_Request();
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $init_Serializer();
     final _rawResponse = await _i1.ResponseConverter<LoginFlowV2, void>(_serializer).convert(_response);
@@ -982,7 +978,7 @@ class $CollaborationResourcesClient {
       filter: filter,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $searchCollections_Serializer();
     final _rawResponse =
@@ -1080,7 +1076,7 @@ class $CollaborationResourcesClient {
       collectionId: collectionId,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $listCollection_Serializer();
     final _rawResponse =
@@ -1187,7 +1183,7 @@ class $CollaborationResourcesClient {
       collectionId: collectionId,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $renameCollection_Serializer();
     final _rawResponse =
@@ -1303,7 +1299,7 @@ class $CollaborationResourcesClient {
       collectionId: collectionId,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $addResource_Serializer();
     final _rawResponse =
@@ -1419,7 +1415,7 @@ class $CollaborationResourcesClient {
       collectionId: collectionId,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $removeResource_Serializer();
     final _rawResponse =
@@ -1525,7 +1521,7 @@ class $CollaborationResourcesClient {
       resourceId: resourceId,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $getCollectionsByResource_Serializer();
     final _rawResponse =
@@ -1644,7 +1640,7 @@ class $CollaborationResourcesClient {
       baseResourceId: baseResourceId,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $createCollectionOnResource_Serializer();
     final _rawResponse =
@@ -1747,7 +1743,7 @@ class $GuestAvatarClient {
       guestName: guestName,
       size: size,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $getAvatarDark_Serializer();
     final _rawResponse = await _i1.ResponseConverter<Uint8List, void>(_serializer).convert(_response);
@@ -1849,7 +1845,7 @@ class $GuestAvatarClient {
       size: size,
       darkTheme: darkTheme,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $getAvatar_Serializer();
     final _rawResponse = await _i1.ResponseConverter<Uint8List, void>(_serializer).convert(_response);
@@ -1949,7 +1945,7 @@ class $HoverCardClient {
       userId: userId,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $getUser_Serializer();
     final _rawResponse =
@@ -2053,7 +2049,7 @@ class $NavigationClient {
       absolute: absolute,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $getAppsNavigation_Serializer();
     final _rawResponse =
@@ -2151,7 +2147,7 @@ class $NavigationClient {
       absolute: absolute,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $getSettingsNavigation_Serializer();
     final _rawResponse =
@@ -2228,7 +2224,7 @@ class $OcmClient {
   ///  * [$discovery_Serializer] for a converter to parse the `Response` from an executed request.
   Future<_i1.DynamiteResponse<OcmDiscoveryResponseApplicationJson, OcmOcmDiscoveryHeaders>> discovery() async {
     final _request = $discovery_Request();
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $discovery_Serializer();
     final _rawResponse =
@@ -2316,7 +2312,7 @@ class $OcsClient {
     final _request = $getCapabilities_Request(
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $getCapabilities_Serializer();
     final _rawResponse =
@@ -2472,7 +2468,7 @@ class $PreviewClient {
       mode: mode,
       mimeFallback: mimeFallback,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $getPreviewByFileId_Serializer();
     final _rawResponse = await _i1.ResponseConverter<Uint8List, void>(_serializer).convert(_response);
@@ -2619,7 +2615,7 @@ class $PreviewClient {
       mode: mode,
       mimeFallback: mimeFallback,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $getPreview_Serializer();
     final _rawResponse = await _i1.ResponseConverter<Uint8List, void>(_serializer).convert(_response);
@@ -2740,7 +2736,7 @@ class $ProfileApiClient {
       targetUserId: targetUserId,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $setVisibility_Serializer();
     final _rawResponse = await _i1.ResponseConverter<ProfileApiSetVisibilityResponseApplicationJson, void>(_serializer)
@@ -2826,7 +2822,7 @@ class $ReferenceClient {
     final _request = $preview_Request(
       referenceId: referenceId,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $preview_Serializer();
     final _rawResponse = await _i1.ResponseConverter<Uint8List, void>(_serializer).convert(_response);
@@ -2925,7 +2921,7 @@ class $ReferenceApiClient {
       reference: reference,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $resolveOne_Serializer();
     final _rawResponse = await _i1.ResponseConverter<ReferenceApiResolveOneResponseApplicationJson, void>(_serializer)
@@ -3028,7 +3024,7 @@ class $ReferenceApiClient {
       limit: limit,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $resolve_Serializer();
     final _rawResponse =
@@ -3139,7 +3135,7 @@ class $ReferenceApiClient {
       limit: limit,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $extract_Serializer();
     final _rawResponse =
@@ -3221,7 +3217,7 @@ class $ReferenceApiClient {
     final _request = $getProvidersInfo_Request(
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $getProvidersInfo_Serializer();
     final _rawResponse =
@@ -3323,7 +3319,7 @@ class $ReferenceApiClient {
       timestamp: timestamp,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $touchProvider_Serializer();
     final _rawResponse =
@@ -3411,7 +3407,7 @@ class $TextProcessingApiClient {
     final _request = $taskTypes_Request(
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $taskTypes_Serializer();
     final _rawResponse =
@@ -3535,7 +3531,7 @@ class $TextProcessingApiClient {
       identifier: identifier,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $schedule_Serializer();
     final _rawResponse =
@@ -3631,7 +3627,7 @@ class $TextProcessingApiClient {
       id: id,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $getTask_Serializer();
     final _rawResponse = await _i1.ResponseConverter<TextProcessingApiGetTaskResponseApplicationJson, void>(_serializer)
@@ -3728,7 +3724,7 @@ class $TextProcessingApiClient {
       id: id,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $deleteTask_Serializer();
     final _rawResponse =
@@ -3832,7 +3828,7 @@ class $TextProcessingApiClient {
       identifier: identifier,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $listTasksByApp_Serializer();
     final _rawResponse =
@@ -3920,7 +3916,7 @@ class $TextToImageApiClient {
     final _request = $isAvailable_Request(
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $isAvailable_Serializer();
     final _rawResponse =
@@ -4043,7 +4039,7 @@ class $TextToImageApiClient {
       numberOfImages: numberOfImages,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $schedule_Serializer();
     final _rawResponse = await _i1.ResponseConverter<TextToImageApiScheduleResponseApplicationJson, void>(_serializer)
@@ -4138,7 +4134,7 @@ class $TextToImageApiClient {
       id: id,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $getTask_Serializer();
     final _rawResponse =
@@ -4235,7 +4231,7 @@ class $TextToImageApiClient {
       id: id,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $deleteTask_Serializer();
     final _rawResponse = await _i1.ResponseConverter<TextToImageApiDeleteTaskResponseApplicationJson, void>(_serializer)
@@ -4337,7 +4333,7 @@ class $TextToImageApiClient {
       index: index,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $getImage_Serializer();
     final _rawResponse = await _i1.ResponseConverter<Uint8List, void>(_serializer).convert(_response);
@@ -4439,7 +4435,7 @@ class $TextToImageApiClient {
       identifier: identifier,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $listTasksByApp_Serializer();
     final _rawResponse =
@@ -4527,7 +4523,7 @@ class $TranslationApiClient {
     final _request = $languages_Request(
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $languages_Serializer();
     final _rawResponse = await _i1.ResponseConverter<TranslationApiLanguagesResponseApplicationJson, void>(_serializer)
@@ -4641,7 +4637,7 @@ class $TranslationApiClient {
       fromLanguage: fromLanguage,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $translate_Serializer();
     final _rawResponse = await _i1.ResponseConverter<TranslationApiTranslateResponseApplicationJson, void>(_serializer)
@@ -4742,7 +4738,7 @@ class $UnifiedSearchClient {
       from: from,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $getProviders_Serializer();
     final _rawResponse =
@@ -4886,7 +4882,7 @@ class $UnifiedSearchClient {
       from: from,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $search_Serializer();
     final _rawResponse =
@@ -4974,7 +4970,7 @@ class $WhatsNewClient {
     final _request = $$get_Request(
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $$get_Serializer();
     final _rawResponse =
@@ -5068,7 +5064,7 @@ class $WhatsNewClient {
       version: version,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $dismiss_Serializer();
     final _rawResponse =
@@ -5154,7 +5150,7 @@ class $WipeClient {
     final _request = $checkWipe_Request(
       token: token,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $checkWipe_Serializer();
     final _rawResponse =
@@ -5233,7 +5229,7 @@ class $WipeClient {
     final _request = $wipeDone_Request(
       token: token,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $wipeDone_Serializer();
     final _rawResponse = await _i1.ResponseConverter<JsonObject, void>(_serializer).convert(_response);
