+++ packages/nextcloud/lib/src/api/dashboard.openapi.dart
@@ -34,9 +34,7 @@ class $Client extends _i1.DynamiteClient {
   /// Creates a new `DynamiteClient` for untagged requests.
   $Client(
     super.baseURL, {
-    super.baseHeaders,
     super.httpClient,
-    super.cookieJar,
     super.authentications,
   });
 
@@ -44,9 +42,7 @@ class $Client extends _i1.DynamiteClient {
   $Client.fromClient(_i1.DynamiteClient client)
       : super(
           client.baseURL,
-          baseHeaders: client.baseHeaders,
           httpClient: client.httpClient,
-          cookieJar: client.cookieJar,
           authentications: client.authentications,
         );
 
@@ -133,7 +129,7 @@ class $DashboardApiClient {
     final _request = $getWidgets_Request(
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $getWidgets_Serializer();
     final _rawResponse = await _i1.ResponseConverter<DashboardApiGetWidgetsResponseApplicationJson, void>(_serializer)
@@ -250,7 +246,7 @@ class $DashboardApiClient {
       widgets: widgets,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $getWidgetItems_Serializer();
     final _rawResponse =
@@ -372,7 +368,7 @@ class $DashboardApiClient {
       widgets: widgets,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $getWidgetItemsV2_Serializer();
     final _rawResponse =
