+++ packages/nextcloud/lib/src/api/dav.openapi.dart
@@ -33,9 +33,7 @@ class $Client extends _i1.DynamiteClient {
   /// Creates a new `DynamiteClient` for untagged requests.
   $Client(
     super.baseURL, {
-    super.baseHeaders,
     super.httpClient,
-    super.cookieJar,
     super.authentications,
   });
 
@@ -43,9 +41,7 @@ class $Client extends _i1.DynamiteClient {
   $Client.fromClient(_i1.DynamiteClient client)
       : super(
           client.baseURL,
-          baseHeaders: client.baseHeaders,
           httpClient: client.httpClient,
-          cookieJar: client.cookieJar,
           authentications: client.authentications,
         );
 
@@ -158,7 +154,7 @@ class $DirectClient {
       expirationTime: expirationTime,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $getUrl_Serializer();
     final _rawResponse =
@@ -261,7 +257,7 @@ class $OutOfOfficeClient {
       userId: userId,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $getCurrentOutOfOfficeData_Serializer();
     final _rawResponse =
@@ -357,7 +353,7 @@ class $OutOfOfficeClient {
       userId: userId,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $getOutOfOffice_Serializer();
     final _rawResponse =
@@ -489,7 +485,7 @@ class $OutOfOfficeClient {
       userId: userId,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $setOutOfOffice_Serializer();
     final _rawResponse =
@@ -585,7 +581,7 @@ class $OutOfOfficeClient {
       userId: userId,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $clearOutOfOffice_Serializer();
     final _rawResponse =
