+++ packages/nextcloud/lib/src/api/files.openapi.dart
@@ -36,9 +36,7 @@ class $Client extends _i1.DynamiteClient {
   /// Creates a new `DynamiteClient` for untagged requests.
   $Client(
     super.baseURL, {
-    super.baseHeaders,
     super.httpClient,
-    super.cookieJar,
     super.authentications,
   });
 
@@ -46,9 +44,7 @@ class $Client extends _i1.DynamiteClient {
   $Client.fromClient(_i1.DynamiteClient client)
       : super(
           client.baseURL,
-          baseHeaders: client.baseHeaders,
           httpClient: client.httpClient,
-          cookieJar: client.cookieJar,
           authentications: client.authentications,
         );
 
@@ -169,7 +165,7 @@ class $ApiClient {
       y: y,
       file: file,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $getThumbnail_Serializer();
     final _rawResponse = await _i1.ResponseConverter<Uint8List, void>(_serializer).convert(_response);
@@ -254,7 +250,7 @@ class $DirectEditingClient {
     final _request = $info_Request(
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $info_Serializer();
     final _rawResponse =
@@ -358,7 +354,7 @@ class $DirectEditingClient {
       creatorId: creatorId,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $templates_Serializer();
     final _rawResponse = await _i1.ResponseConverter<DirectEditingTemplatesResponseApplicationJson, void>(_serializer)
@@ -471,7 +467,7 @@ class $DirectEditingClient {
       fileId: fileId,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $open_Serializer();
     final _rawResponse =
@@ -594,7 +590,7 @@ class $DirectEditingClient {
       templateId: templateId,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $create_Serializer();
     final _rawResponse =
@@ -696,7 +692,7 @@ class $OpenLocalEditorClient {
       path: path,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $create_Serializer();
     final _rawResponse =
@@ -799,7 +795,7 @@ class $OpenLocalEditorClient {
       token: token,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $validate_Serializer();
     final _rawResponse = await _i1.ResponseConverter<OpenLocalEditorValidateResponseApplicationJson, void>(_serializer)
@@ -885,7 +881,7 @@ class $TemplateClient {
     final _request = $list_Request(
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $list_Serializer();
     final _rawResponse =
@@ -999,7 +995,7 @@ class $TemplateClient {
       templateType: templateType,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $create_Serializer();
     final _rawResponse =
@@ -1107,7 +1103,7 @@ class $TemplateClient {
       copySystemTemplates: copySystemTemplates,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $path_Serializer();
     final _rawResponse =
@@ -1220,7 +1216,7 @@ class $TransferOwnershipClient {
       path: path,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $transfer_Serializer();
     final _rawResponse =
@@ -1318,7 +1314,7 @@ class $TransferOwnershipClient {
       id: id,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $accept_Serializer();
     final _rawResponse = await _i1.ResponseConverter<TransferOwnershipAcceptResponseApplicationJson, void>(_serializer)
@@ -1415,7 +1411,7 @@ class $TransferOwnershipClient {
       id: id,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $reject_Serializer();
     final _rawResponse = await _i1.ResponseConverter<TransferOwnershipRejectResponseApplicationJson, void>(_serializer)
