+++ packages/nextcloud/lib/src/api/notifications.openapi.dart
@@ -34,9 +34,7 @@ class $Client extends _i1.DynamiteClient {
   /// Creates a new `DynamiteClient` for untagged requests.
   $Client(
     super.baseURL, {
-    super.baseHeaders,
     super.httpClient,
-    super.cookieJar,
     super.authentications,
   });
 
@@ -44,9 +42,7 @@ class $Client extends _i1.DynamiteClient {
   $Client.fromClient(_i1.DynamiteClient client)
       : super(
           client.baseURL,
-          baseHeaders: client.baseHeaders,
           httpClient: client.httpClient,
-          cookieJar: client.cookieJar,
           authentications: client.authentications,
         );
 
@@ -189,7 +185,7 @@ class $ApiClient {
       apiVersion: apiVersion,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $generateNotification_Serializer();
     final _rawResponse = await _i1.ResponseConverter<ApiGenerateNotificationResponseApplicationJson, void>(_serializer)
@@ -296,7 +292,7 @@ class $EndpointClient {
       apiVersion: apiVersion,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $listNotifications_Serializer();
     final _rawResponse = await _i1.ResponseConverter<EndpointListNotificationsResponseApplicationJson,
@@ -396,7 +392,7 @@ class $EndpointClient {
       apiVersion: apiVersion,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $deleteAllNotifications_Serializer();
     final _rawResponse =
@@ -503,7 +499,7 @@ class $EndpointClient {
       apiVersion: apiVersion,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $getNotification_Serializer();
     final _rawResponse = await _i1.ResponseConverter<EndpointGetNotificationResponseApplicationJson, void>(_serializer)
@@ -611,7 +607,7 @@ class $EndpointClient {
       apiVersion: apiVersion,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $deleteNotification_Serializer();
     final _rawResponse =
@@ -718,7 +714,7 @@ class $EndpointClient {
       apiVersion: apiVersion,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $confirmIdsForUser_Serializer();
     final _rawResponse =
@@ -853,7 +849,7 @@ class $PushClient {
       apiVersion: apiVersion,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $registerDevice_Serializer();
     final _rawResponse =
@@ -954,7 +950,7 @@ class $PushClient {
       apiVersion: apiVersion,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $removeDevice_Serializer();
     final _rawResponse =
@@ -1082,7 +1078,7 @@ class $SettingsClient {
       apiVersion: apiVersion,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $personal_Serializer();
     final _rawResponse =
@@ -1205,7 +1201,7 @@ class $SettingsClient {
       apiVersion: apiVersion,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $admin_Serializer();
     final _rawResponse =
