+++ packages/nextcloud/lib/src/api/theming.openapi.dart
@@ -36,9 +36,7 @@ class $Client extends _i1.DynamiteClient {
   /// Creates a new `DynamiteClient` for untagged requests.
   $Client(
     super.baseURL, {
-    super.baseHeaders,
     super.httpClient,
-    super.cookieJar,
     super.authentications,
   });
 
@@ -46,9 +44,7 @@ class $Client extends _i1.DynamiteClient {
   $Client.fromClient(_i1.DynamiteClient client)
       : super(
           client.baseURL,
-          baseHeaders: client.baseHeaders,
           httpClient: client.httpClient,
-          cookieJar: client.cookieJar,
           authentications: client.authentications,
         );
 
@@ -141,7 +137,7 @@ class $IconClient {
     final _request = $getFavicon_Request(
       app: app,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $getFavicon_Serializer();
     final _rawResponse = await _i1.ResponseConverter<Uint8List, void>(_serializer).convert(_response);
@@ -222,7 +218,7 @@ class $IconClient {
     final _request = $getTouchIcon_Request(
       app: app,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $getTouchIcon_Serializer();
     final _rawResponse = await _i1.ResponseConverter<Uint8List, void>(_serializer).convert(_response);
@@ -319,7 +315,7 @@ class $IconClient {
       app: app,
       image: image,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $getThemedIcon_Serializer();
     final _rawResponse = await _i1.ResponseConverter<Uint8List, void>(_serializer).convert(_response);
@@ -432,7 +428,7 @@ class $ThemingClient {
       plain: plain,
       withCustomCss: withCustomCss,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $getThemeStylesheet_Serializer();
     final _rawResponse = await _i1.ResponseConverter<String, void>(_serializer).convert(_response);
@@ -525,7 +521,7 @@ class $ThemingClient {
       key: key,
       useSvg: useSvg,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $getImage_Serializer();
     final _rawResponse = await _i1.ResponseConverter<Uint8List, void>(_serializer).convert(_response);
@@ -605,7 +601,7 @@ class $ThemingClient {
     final _request = $getManifest_Request(
       app: app,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $getManifest_Serializer();
     final _rawResponse =
@@ -681,7 +677,7 @@ class $UserThemeClient {
   ///  * [$getBackground_Serializer] for a converter to parse the `Response` from an executed request.
   Future<_i1.DynamiteResponse<Uint8List, void>> getBackground() async {
     final _request = $getBackground_Request();
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $getBackground_Serializer();
     final _rawResponse = await _i1.ResponseConverter<Uint8List, void>(_serializer).convert(_response);
@@ -777,7 +773,7 @@ class $UserThemeClient {
       themeId: themeId,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $enableTheme_Serializer();
     final _rawResponse =
@@ -874,7 +870,7 @@ class $UserThemeClient {
       themeId: themeId,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $disableTheme_Serializer();
     final _rawResponse =
