+++ packages/nextcloud/lib/src/api/weather_status.openapi.dart
@@ -33,9 +33,7 @@ class $Client extends _i1.DynamiteClient {
   /// Creates a new `DynamiteClient` for untagged requests.
   $Client(
     super.baseURL, {
-    super.baseHeaders,
     super.httpClient,
-    super.cookieJar,
     super.authentications,
   });
 
@@ -43,9 +41,7 @@ class $Client extends _i1.DynamiteClient {
   $Client.fromClient(_i1.DynamiteClient client)
       : super(
           client.baseURL,
-          baseHeaders: client.baseHeaders,
           httpClient: client.httpClient,
-          cookieJar: client.cookieJar,
           authentications: client.authentications,
         );
 
@@ -143,7 +139,7 @@ class $WeatherStatusClient {
       mode: mode,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $setMode_Serializer();
     final _rawResponse =
@@ -225,7 +221,7 @@ class $WeatherStatusClient {
     final _request = $usePersonalAddress_Request(
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $usePersonalAddress_Serializer();
     final _rawResponse =
@@ -308,7 +304,7 @@ class $WeatherStatusClient {
     final _request = $getLocation_Request(
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $getLocation_Serializer();
     final _rawResponse = await _i1.ResponseConverter<WeatherStatusGetLocationResponseApplicationJson, void>(_serializer)
@@ -418,7 +414,7 @@ class $WeatherStatusClient {
       lon: lon,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $setLocation_Serializer();
     final _rawResponse = await _i1.ResponseConverter<WeatherStatusSetLocationResponseApplicationJson, void>(_serializer)
@@ -502,7 +498,7 @@ class $WeatherStatusClient {
     final _request = $getForecast_Request(
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $getForecast_Serializer();
     final _rawResponse = await _i1.ResponseConverter<WeatherStatusGetForecastResponseApplicationJson, void>(_serializer)
@@ -584,7 +580,7 @@ class $WeatherStatusClient {
     final _request = $getFavorites_Request(
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $getFavorites_Serializer();
     final _rawResponse =
@@ -680,7 +676,7 @@ class $WeatherStatusClient {
       favorites: favorites,
       oCSAPIRequest: oCSAPIRequest,
     );
-    final _response = await _rootClient.send(_request);
+    final _response = await _rootClient.httpClient.send(_request);
 
     final _serializer = $setFavorites_Serializer();
     final _rawResponse =
