+++ packages/nextcloud/pubspec.yaml
@@ -15,6 +15,7 @@ dependencies:
   built_collection: ^5.0.0
   built_value: ^8.9.0
   collection: ^1.0.0
+  cookie_jar: ^4.0.7
   crypto: ^3.0.0
   crypton: ^2.0.0
   dynamite_runtime: ^0.2.0
